v1.0.7.0 (Full Release)

+++ Barotrauma/BarotraumaClient/ClientSource/Characters/CharacterInfo.cs
@@ -578,7 +578,7 @@ namespace Barotrauma
             ch.SetPersonalityTrait();
             ch.Job?.OverrideSkills(skillLevels);
 
-            ch.ExperiencePoints = inc.ReadUInt16();
+            ch.ExperiencePoints = inc.ReadInt32();
             ch.AdditionalTalentPoints = inc.ReadRangedInteger(0, MaxAdditionalTalentPoints);
             return ch;
         }
