+++ Barotrauma/BarotraumaServer/ServerSource/Characters/CharacterInfo.cs
@@ -91,7 +91,7 @@ namespace Barotrauma
                 msg.WriteByte((byte)0);
             }
 
-            msg.WriteUInt16((ushort)ExperiencePoints);
+            msg.WriteInt32(ExperiencePoints);
             msg.WriteRangedInteger(AdditionalTalentPoints, 0, MaxAdditionalTalentPoints);
         }
     }
